requirements.txt
discord
tinyDB
python-dotenv
